Fixed some prop type errors
